requirements.txt
groq
Pillow
pydantic
python-dotenv

